chore: fix errors caused by merge main.
