+++ tools/tdgpt/taosanalytics/app.py
@@ -1,7 +1,6 @@
 # encoding:utf-8
 # pylint: disable=c0103
 """the main route definition for restful service"""
-import os
 import os.path
 import sys
 
@@ -13,7 +12,7 @@ from taosanalytics.handlers.imputation import handle_imputation
 from taosanalytics.handlers.anomaly import handle_anomaly
 from taosanalytics.handlers.forecast import handle_forecast
 from taosanalytics.handlers.correlation import handle_correlation
-from taosanalytics.handlers.misc import handle_batch
+from taosanalytics.handlers.misc import do_profile_search, handle_batch, handle_pearsonr
 
 from taosanalytics.conf import Configure
 from taosanalytics.log import AppLogger
@@ -44,7 +43,6 @@ def _init_app():
 # Create Flask app
 app = Flask(__name__)
 app.config["PROPAGATE_EXCEPTIONS"] = True
-
 
 @app.route("/")
 def index():
@@ -122,154 +120,17 @@ def undeploy_model():
     return do_handle_undeploy_model(request)
 
 
-def handle_pearsonr(request, api_version):
-    """
-    Execute pearsonr correlation logic.
-
-    :param request: Flask request object
-    :param api_version: API version to determine the specific implementation of Pearson correlation
-    :return: dict with correlation result or error information
-    """
-    try:
-        # check for rows limitation to reduce the process time
-        req_json, payload, options, data_index, _ = do_check_before_exec(request, False)
-    except ValueError as e:
-        msg = str(e)
-        if msg == SINGLE_COLUMN_ERROR_MSG:
-            msg = 'a second data column is required for pearsonr'
-        return {"msg": msg, "rows": -1}
-    except Exception as e:
-        return {"msg": str(e), "rows": -1}
-
-    if api_version != 'v1':
-        app_logger.log_inst.error('unsupported API version: %s', api_version)
-        return {"msg": f"unsupported API version: {api_version}", "rows": -1}
-
-    try:
-        second_list = get_more_data_list(payload, req_json["schema"])
-        if second_list is None:
-            return {"msg": "a second data column is required for pearsonr", "rows": -1}
-        
-        correlation, p_value = pearsonr(payload[data_index], second_list)
-        if not np.isfinite(correlation):
-            correlation = 0.0
-            p_value = 1.0
-
-        correlation = float(correlation)
-        p_value = float(p_value)
-
-        app_logger.log_inst.debug(f"pearsonr correlation: {correlation}, p value: {p_value}")
-        res = {"option": options, "rows": 1, "correlation_coefficient": correlation, "p_value": p_value}
-
-        return res
-    except Exception as e:
-        app_logger.log_inst.error('pearsonr correlation failed, %s', str(e))
-        return {"msg": str(e), "rows": -1}
-
-def do_profile_search(request, api_version):
-    """
-    Execute profile search logic.
-
-    :param request: Flask request object
-    :param api_version: API version to determine the specific implementation of profile search
-    :return: dict with profile search result or error information
-
-    Request body example:
-    Supported normalization values: "min-max", "z-score", "centering", "none".
-    Supported algo types: "dtw", "cosine".
-    For "dtw", algo.params may include:
-    - "radius": search radius for fastdtw
-    - "min_window": minimum ts window size for profile search
-    - "max_window": maximum ts window size for profile search
-    Result selection notes:
-    - Return the top N similar profiles with "num".
-    - Or return all profiles with distance below the threshold when using dtw.
-    - Or return all profiles with similarity above the threshold when using cosine similarity.
-    - "num" and "threshold" cannot be set at the same time.
-    - Threshold-based results are capped at 500 matches.
-    target_data.ts may be either:
-    - a unix timestamp list, such as [1, 2, 3, 4, 5, 6]
-    - a ts window list, such as [[1, 5], [2, 6]]
-    {
-        "normalization": "z-score",
-        "algo": {
-            "type": "dtw",
-            "params": {
-                "radius": 5,
-                "min_window": 5,
-                "max_window": 20
-            }
-        },
-        "result": {
-            "num": 3
-        },
-        "source_data": [1, 2, 3, 4, 5],
-        "target_data": {
-            "ts": [1, 2, 3, 4, 5, 6],
-            "data": [
-                [1, 2, 3, 4, 5]
-            ]
-        }
-    }
-    Response example:
-    metric_type is either "dtw_distance" or "cosine_similarity".
-    Sort rule:
-    - "dtw_distance": smaller "criteria" means more similar (ascending)
-    - "cosine_similarity": larger "criteria" means more similar (descending)
-    In each match, "num" is the number of data points in the matched window.
-    {
-        "rows": 3,
-        "metric_type": "dtw_distance",
-        "matches": [
-            {
-                "criteria": 0.12,
-                "ts_window": [1, 5],
-                "num": 7
-            },
-            {
-                "criteria": 0.21,
-                "ts_window": [2, 6],
-                "num": 5
-            },
-            {
-                "criteria": 0.35,
-                "ts_window": [3, 7],
-                "num": 4
-            }
-        ]
-    }
-
-    """
-    if api_version != 'v1':
-        app_logger.log_inst.error('unsupported API version: %s', api_version)
-        return {"msg": f"unsupported API version: {api_version}", "rows": -1}
-
-    try:
-        req_json = do_initial_check(request)
-    except Exception as e:
-        return {"msg": str(e), "rows": -1}
-
-    try:
-        result = do_profile_search_impl(req_json)
-        app_logger.log_inst.debug("profile-search result: %s", result)
-        return result
-
-    except Exception as e:
-        app_logger.log_inst.error('profile search failed, %s', str(e))
-        return {"msg": str(e), "rows": -1}
-
-
 @app.route('/api/v1/analysis/pearsonr', methods=['POST'])
 def handle_pearsonr_req():
     """handle the pearsonr correlation request """
-    app_logger.log_inst.info('recv pearsonr correlation request from %s', request.remote_addr)
+    AppLogger.info('recv pearsonr correlation request from %s', request.remote_addr)
     return handle_pearsonr(request, api_version='v1')
 
 
 @app.route('/api/v1/analysis/profile-search', methods=['POST'])
 def handle_profile_search_req():
     """handle the profile search request """
-    app_logger.log_inst.info('recv profile search request from %s', request.remote_addr)
+    AppLogger.info('recv profile search request from %s', request.remote_addr)
     return do_profile_search(request, api_version='v1')
